+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
 
+
+## [2.52.0] - 2022-05-13
+### Changed
+- tracking of featured piece clicks from Livewire to Vue
+- FeaturedPiece supports rich formatting
+
 ## [2.51.2] - 2022-05-11
 ### Fixed
 - constructing importers